+++ Sources/AWSCognitoAuthentication/Request.swift
@@ -1,12 +1,5 @@
-import CognitoIdentityProvider
-import NIO
+import AWSCognitoAuthenticationKit
 import Vapor
-
-/// Protocol for objects that encompass both an eventloop and context data to be used by Cognito
-public protocol AWSCognitoEventLoopWithContext {
-    var eventLoop: EventLoop { get }
-    var cognitoContextData: CognitoIdentityProvider.ContextDataType? { get }
-}
 
 /// extend Vapor Request to provide Cognito context
 extension Request: AWSCognitoEventLoopWithContext {
