+++ Sources/AWSCognitoAuthenticationKit/Authenticatable+SRP.swift
@@ -166,3 +166,13 @@ class SRP<H: HashFunction> {
     }
 }
 
+extension Array where Element: FixedWidthInteger {
+    static func random(count: Int) -> [Element] {
+        var array = self.init()
+        for _ in 0..<count {
+            array.append(.random(in: Element.min..<Element.max))
+        }
+        return array
+    }
+}
+
