+++ Sources/AWSCognitoAuthenticationKit/Identifiable.swift
